separate feature model testing from mixset testing
