+++ cruise.umple/src/Master.ump
@@ -81,3 +81,8 @@ use util/Util_Code.ump;
 use util/Ant.ump;
 use Builder_Code.ump;
 use GraphWalking.ump;
+
+
+mixset FeatureModel {
+  require [Mixset];
+}
